fix: extract target checks
