+++ helpers/doesNotNeedPlugin.js
@@ -1,20 +1,14 @@
 const findUp = require('find-up')
 
 // Checks all the cases for which the plugin should do nothing
-const isStaticExportProject = require('./isStaticExportProject')
 const doesSiteUseNextOnNetlify = require('./doesSiteUseNextOnNetlify')
-const hasCorrectNextConfig = require('./hasCorrectNextConfig')
+const isStaticExportProject = require('./isStaticExportProject')
 
-const doesNotNeedPlugin = async ({ netlifyConfig, packageJson, failBuild }) => {
+const doesNotNeedPlugin = ({ netlifyConfig, packageJson }) => {
   const { build } = netlifyConfig
-  const { name, scripts = {} } = packageJson
-  const nextConfigPath = await findUp('next.config.js')
+  const { scripts = {} } = packageJson
 
-  return (
-    isStaticExportProject({ build, scripts }) ||
-    doesSiteUseNextOnNetlify({ packageJson }) ||
-    !(await hasCorrectNextConfig({ nextConfigPath, failBuild }))
-  )
+  return isStaticExportProject({ build, scripts }) || doesSiteUseNextOnNetlify({ packageJson })
 }
 
 module.exports = doesNotNeedPlugin
