+++ index.js
@@ -1,8 +1,3 @@
-const fs = require('fs')
-const path = require('path')
-const util = require('util')
-
-const findUp = require('find-up')
 const makeDir = require('make-dir')
 
 const { restoreCache, saveCache } = require('./helpers/cacheBuild')
@@ -10,9 +5,8 @@ const copyUnstableIncludedDirs = require('./helpers/copyUnstableIncludedDirs')
 const doesNotNeedPlugin = require('./helpers/doesNotNeedPlugin')
 const getNextConfig = require('./helpers/getNextConfig')
 const validateNextUsage = require('./helpers/validateNextUsage')
+const verifyBuildTarget = require('./helpers/verifyBuildTarget')
 const nextOnNetlify = require('./src/index.js')
-
-const pWriteFile = util.promisify(fs.writeFile)
 
 // * Helpful Plugin Context *
 // - Between the prebuild and build steps, the project's build command is run
@@ -28,8 +22,13 @@ module.exports = {
     if (hasNoPackageJson) {
       return failBuild('Could not find a package.json for this project')
     }
+
+    if (doesNotNeedPlugin({ netlifyConfig, packageJson, failBuild })) {
+      return
+    }
+
     // Populates the correct config if needed
-    await doesNotNeedPlugin({ netlifyConfig, packageJson, failBuild })
+    await verifyBuildTarget({ netlifyConfig, packageJson, failBuild })
 
     // Because we memoize nextConfig, we need to do this after the write file
     const nextConfig = await getNextConfig(utils.failBuild)
@@ -51,7 +50,7 @@ module.exports = {
   }) {
     const { failBuild } = utils.build
 
-    if (await doesNotNeedPlugin({ netlifyConfig, packageJson, failBuild })) {
+    if (doesNotNeedPlugin({ netlifyConfig, packageJson, failBuild })) {
       return
     }
 
@@ -63,7 +62,7 @@ module.exports = {
   },
 
   async onPostBuild({ netlifyConfig, packageJson, constants: { FUNCTIONS_DIST }, utils }) {
-    if (await doesNotNeedPlugin({ netlifyConfig, packageJson, utils })) {
+    if (doesNotNeedPlugin({ netlifyConfig, packageJson, utils })) {
       return
     }
